requirements.txt
python-dotenv==0.10.1
Django==2.2.7
djangorestframework==3.10.3
psycopg2==2.7.3.2
ipython==7.9.0
opencv-python==4.1.1.26
numpy==1.17.3
django-jet==1.0.8
django-filter==2.2.0
django-health-check==3.11.1